requirements.txt
python-dotenv
python-telegram-bot==13.13
requests==2.28.1
urllib3==1.26.15
openai